+++ .eslintrc.js
@@ -10,7 +10,7 @@ module.exports = {
     // List of polyfills for `eslint-plugin-compat` check
     // To know how to add in case you have a new one to add, please check
     // https://github.com/amilajack/eslint-plugin-compat/wiki/Adding-polyfills-(v2)
-    polyfills: [],
+    polyfills: ['Promise.resolve'],
   },
   parserOptions: {
     ecmaVersion: 2018,