+++ src/index.ts
@@ -2,5 +2,7 @@ import './user-timing-api-resolver';
 
 export * from './marks';
 export * from './entries';
+export * from './profiler';
+export * from './is-nodejs-env';
 export * from './is-user-timing-api-supported';
 export * from './is-performance-observable-supported';