+++ src/user-timing-api-resolver.ts
@@ -4,10 +4,10 @@ if (isNodeJSEnv && !(global as any).PerformanceObserver && !(global as any).perf
   /**
    * Requires a module which is protected against bundler minification.
    *
-   * @param request The module path to resolve
+   * @param pkg The module path to resolve
    */
-  const dynamicRequire = (mod: { require: (r: string) => any }, request: string): any => {
-    return mod.require(request);
+  const dynamicRequire = (mod: { require: (r: string) => any }, pkg: string): any => {
+    return mod.require(pkg);
   };
 
   try {
@@ -17,6 +17,6 @@ if (isNodeJSEnv && !(global as any).PerformanceObserver && !(global as any).perf
     (global as any).PerformanceObserver = PerformanceObserver;
     (global as any).performance = performance;
   } catch (error) {
-    throw new Error(`Your NodeJS application doesn't support perf_hooks. ${error}`);
+    throw new Error(`Your NodeJS application doesn't support 'perf_hooks'. ${error}`);
   }
 }